backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
alembic==1.13.1
psycopg[binary]==3.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.1
itsdangerous==2.1.2
fastapi-mail==1.4.1
